refactor(plugins): give the plugin capability surface its own module
This is the security-review surface: the verbs a plugin can actually reach —
ctx.messages, ctx.engine, ctx.net, ctx.storage, ctx.conversations, ctx.handover,
ctx.mappings — and the gates in front of every one of them. A plugin supplies
its own sessionId, so assertSessionAllowed (the static manifest boundary) and
the operator-activation check ARE the boundary, not a convenience.

It sat inside a loader that is otherwise about directory discovery, lifecycle
and worker plumbing, so reviewing the boundary meant reading 1640 lines to find
330. PluginCapabilityContext now holds the seven gate helpers, createPluginContext
and the hook AsyncLocalStorage. It holds no plugin registry and no worker hosts —
everything it needs comes from the PluginInstance it is handed.

The constructor SIGNATURE is untouched: the facade is built in the body, as
PluginHostServices already is. All 33 positional spec constructions and the three
subclasses overriding createSandboxHost compile and pass unchanged.

Two things this surfaced that are worth recording:

The loader's logger is passed in rather than a second one created. A plugin's
ctx.logger lines carry that logger's context, which is operator-visible; a fresh
createLogger('PluginCapabilityContext') silently retagged every plugin log line,
and the sandbox log-relay tests caught it.

dispatchConversationMedia moved too. Leaving it behind would have made the new
module import the loader while the loader imports it — a require cycle, in a file
that already pays a lazy-require tax to avoid exactly that. The loader no longer
references it.

1640 to 1267 lines. Spec changes are pointer-only across four files: three
reach-in helpers now name the object that owns the surface, and one import line
follows dispatchConversationMedia. No assertion, fixture or harness was touched;
307 plugin tests and 4012 overall pass unchanged.
